feature: 피드 상세페이지 추가 및 컴포넌트 리팩토링
