+++ src/app/(pages)/(Home)/p/[feedId]/page.tsx
@@ -1,0 +1,5 @@
+import FeedPage from "@components/pages/FeedPage";
+
+export default function page() {
+  return <FeedPage />;
+}
