+++ src/app/_components/organism/feedDetailBox/FeedDetailBox.style.tsx
