+++ src/app/_components/pages/FeedPage.tsx
@@ -1,0 +1,5 @@
+import FeedTemplate from "../template/feed/FeedTemplate";
+
+export default function FeedPage() {
+  return <FeedTemplate />;
+}
