+++ src/app/_components/template/feed/FeedTemplate.tsx
@@ -1,0 +1,13 @@
+"use client";
+import FeedDetailBox from "@/_components/organism/feedDetailBox/FeedDetailBox";
+import { HomeLayoutStyle } from "@/_styles/common.style";
+
+export default function FeedTemplate() {
+  return (
+    <HomeLayoutStyle.Container>
+      <HomeLayoutStyle.Main>
+        <FeedDetailBox />
+      </HomeLayoutStyle.Main>
+    </HomeLayoutStyle.Container>
+  );
+}
